+++ lib/stylus/railtie.rb
@@ -2,8 +2,7 @@ require 'stylus/tilt'
 module Stylus
   ### Stylus Railtie
   #
-  # `Railtie` responsible for injecting `stylus` inside the
-  # Rails application and the `Sprockets` Asset Pipeline.
+  # `Railtie` that hooks `stylus` inside a Rails application.
   class Railtie < ::Rails::Railtie
 
     config.app_generators.stylesheet_engine :stylus
@@ -12,10 +11,13 @@ module Stylus
       app.assets.register_engine '.styl', Tilt::StylusTemplate
     end
 
-    # Includes the `Rails` asset load path into `stylus` so any
-    # `.styl` file inside it can be imported by the `stylus` API.
+    # After initialization block to inspect the `Sprockets` configuration
+    # And reflect it on the `Stylus` module;
+    # It also includes the `Rails` asset load path into `Stylus` so any
+    # `.styl` file inside it can be imported by the `Stylus` API.
     config.after_initialize do |app|
       Stylus.compress = app.config.assets.compress
+      Stylus.debug = app.config.assets.debug
       Stylus.paths.concat app.assets.paths
     end
   end