+++ spec/spec_helper.rb
@@ -1,3 +1,14 @@
+require 'rails'
+require 'rails/test_help'
+require 'rails/generators/test_case'
+
+require 'active_support/railtie'
+require 'action_controller/railtie'
+require 'sprockets/railtie'
+require 'stylus/railtie'
+
+require 'active_support/core_ext/class/attribute_accessors'
+
 require 'support/helpers'
 require 'support/matchers'
 require 'support/generators/test_case'
@@ -7,7 +18,9 @@ RSpec.configure do |config|
 
   config.after :each do
     Stylus.compress = false
+    Stylus.debug = false
     Stylus.paths = []
     Stylus.plugins.clear
+    Stylus.imports.clear
   end
 end